src/main/java/com/example/ecommerce/WishListApplication.java
package com.example.ecommerce;

import com.example.ecommerce.model.entity.Role;
import com.example.ecommerce.repository.RoleRepository;
import org.springframework.boot.CommandLineRunner;
import org.springframework.boot.SpringApplication;
import org.springframework.boot.autoconfigure.SpringBootApplication;
import org.springframework.context.annotation.Bean;


@SpringBootApplication
public class WishListApplication {
    public static void main(String[] args) {
        SpringApplication.run(WishListApplication.class);
    }
    @Bean
    public CommandLineRunner demo(RoleRepository roleRepo) {
        return (args) -> {
            Role role=new Role();
            role.setName("ROLE_ADMIN");
            roleRepo.save(role);
        };
    }
}